move WAL service implementation into internal/instance package
