+++ internal/instance/start.go
@@ -12,7 +12,6 @@ import (
 	"github.com/cloudnative-pg/cnpg-i/pkg/metrics"
 	"github.com/cloudnative-pg/cnpg-i/pkg/wal"
 	"github.com/dalibo/cnpg-i-pgbackrest/internal/utils"
-	wal_pgbackrest "github.com/dalibo/cnpg-i-pgbackrest/internal/wal"
 	"google.golang.org/grpc"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -30,7 +29,7 @@ type PgbackrestPluginServer struct {
 // Start starts the GRPC service
 func (c *PgbackrestPluginServer) Start(ctx context.Context) error {
 	enrich := func(server *grpc.Server) error {
-		wal.RegisterWALServer(server, &wal_pgbackrest.WALSrvImplementation{
+		wal.RegisterWALServer(server, &WALSrvImplementation{
 			InstanceName: c.InstanceName,
 			Client:       c.Client,
 			PGDataPath:   c.PGDataPath,
