+++ internal/restore/start.go
@@ -10,8 +10,9 @@ import (
 	"github.com/cloudnative-pg/cnpg-i-machinery/pkg/pluginhelper/http"
 	restore "github.com/cloudnative-pg/cnpg-i/pkg/restore/job"
 	"github.com/cloudnative-pg/cnpg-i/pkg/wal"
+
+	"github.com/dalibo/cnpg-i-pgbackrest/internal/instance"
 	"github.com/dalibo/cnpg-i-pgbackrest/internal/utils"
-	wal_pgbackrest "github.com/dalibo/cnpg-i-pgbackrest/internal/wal"
 	"google.golang.org/grpc"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -29,7 +30,7 @@ type CNPGI struct {
 // Start starts the GRPC service
 func (c *CNPGI) Start(ctx context.Context) error {
 	enrich := func(server *grpc.Server) error {
-		wal.RegisterWALServer(server, &wal_pgbackrest.WALSrvImplementation{
+		wal.RegisterWALServer(server, &instance.WALSrvImplementation{
 			InstanceName: c.InstanceName,
 			Client:       c.Client,
 			PGDataPath:   c.PGDataPath,
