--- internal/wal/doc.go
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: 2025 Dalibo <contact@dalibo.com>
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package wal
